refactor: ecsCluster.tsの構造化。
